reverted previous change of variable type in MITHNEOS_to_spectrum
line 816, int to float. Previously changed from float to int, but, as it turns out, not all data in MITHNEOS have the same number of columns. Some 200-300 are missing the 4th column for number of frames used as an average to create the data point. Hence, the column with index -1 is the noise as a string, hence throws an exception when casted to int from base 10.